requirements.txt
biopython>=1.79
pytest>=7.0.0
pytest-cov>=4.0.0
